refactor: 동시성 제어에 대한 통합 테스트 분리
